C API: refactor test support
