+++ tests/unit/libexpr/nix_api_expr.cc
@@ -5,29 +5,11 @@
 #include "nix_api_expr.h"
 #include "nix_api_value.h"
 
-#include "tests/nix_api_store.hh"
+#include "tests/nix_api_expr.hh"
 
 #include <gtest/gtest.h>
 
 namespace nixC {
-
-class nix_api_expr_test : public nix_api_store_test
-{
-public:
-    nix_api_expr_test()
-    {
-        state = nix_state_create(nullptr, nullptr, store);
-        value = nix_alloc_value(nullptr, state);
-    }
-    ~nix_api_expr_test()
-    {
-        nix_gc_decref(nullptr, value);
-        nix_state_free(state);
-    }
-
-    EvalState * state;
-    Value * value;
-};
 
 TEST_F(nix_api_expr_test, nix_expr_eval_from_string)
 {
